refactor(app-chrome): consolidate Header + Footer into AppChrome.tsx
Merges AppHeader.tsx and AppFooter.tsx into AppChrome.tsx as private
functions — no separate files needed for tightly coupled inner components.
